[analyzer] Adjust several visitors
This CL adjusts several visitors to visit `type` rather than
`functionType` of `GenericTypeAlias`. This is needed because
`functionType` is now a view on `type`, and the former will be null
in some cases such that the underlying `type` is incorrectly ignored.

Here are the locations not yet modified in this manner:

Visitors:
  ResolvedAstPrinter
  AstBinaryWriter
  BestPracticesVerifier

Non-visitors:
  TypeAliasSelfReferenceFinder._sanitizeGenericTypeAlias, _Finder.genericTypeAlias
  VarianceBuilder._genericTypeAlias
  NamedTypeBuilder._getRawFunctionType,
  LinkedUnitContext.getGeneticTypeAliasFunction
  _File._buildFileDeclarations
  _LibraryBuilder._addGenericTypeAlias

analysis_server, visitors:
  _LocalBestTypeVisitor

analysis_server, non-visitors:
  InlineTypedef.compute
  DartUnitOutlineComputer._newGenericTypeAliasOutline
  Files: completion_metrics (about 3 locations in here).

It would be great if I could get some feedback on these: Which ones
are most urgent? Will we keep any of them unchanged?

Change-Id: Ifd55c19571d84bb7ec004b65c5f1f22f3b22cb5c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/171586
Commit-Queue: Erik Ernst <eernst@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>