[vm/concurrency] Remove ICData::AddTarget by adding ICData::NewForStaticCall
The `ICData::AddTarget()` is always used directly after allocation of an
ICData (if needed). We'll remove it and add a `ICData::NewForStaticCall()`
which guarantees the ICData is correctly initialized from the start.

This avoids mental overhead of needing to think about whether
`AddTarget()` needs locking or not.

TEST=Refactoring of already tested code.

Issue https://github.com/dart-lang/sdk/issues/36097

Change-Id: I3529dd881887afb281c83b72117bdb9247a8dff0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/173760
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>